Disable 0px -> 0: always keep unit
- 0px will stay 0px
- 0% will stay 0%
- etc.

For Windows Store apps using Internet Explorer, an older standard of
Flexbox is used, where the last param to `flex` must have a unit, not be
unitless: any specific setting of `0px` will be optimised to `0` in
those cases. It's simpler to keep the unit in all cases than
specifically ignore for `flex`.